Add requests in fragments

+++ app/src/main/java/com/davidmendozamartinez/gangame/data/TopGameMapper.kt
@@ -2,13 +2,13 @@ package com.davidmendozamartinez.gangame.data
 
 import com.davidmendozamartinez.gangame.TopGame
 
-object  TopGameMapper {
+object TopGameMapper {
 
     fun fromSdk(topGame: com.davidmendozamartinez.gangame.sdk.TopGame, position: Int): TopGame =
         TopGame(
             topGame.title,
             topGame.owners,
-            topGame.positive / (topGame.positive + topGame.negative) * 100,
+            (topGame.positive.toFloat() / (topGame.positive + topGame.negative) * 100).toInt(),
             topGame.publisher,
             topGame.price / 100F,
             position,
